Move generate_ext_tables to tools, add to deid_runner.sh
